core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "kafka-python>=2.0.2",
    "numba>=0.60.0",
    "onnxruntime>=1.20.1",
    "opencv-contrib-python>=4.10.0.84",
    "opencv-python-headless>=4.10.0.84",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
